+++ src/pull_parser/error/mod.rs
@@ -6,6 +6,8 @@ use std::error;
 use std::fmt;
 use std::io;
 
+use crate::pull_parser::SyntacticPosition;
+
 pub use self::data::{Compression, DataError};
 pub use self::operation::OperationError;
 pub use self::warning::Warning;
@@ -21,35 +23,53 @@ pub type Result<T> = std::result::Result<T, Error>;
 #[derive(Debug)]
 pub struct Error {
     /// The real error.
-    inner: Box<ErrorContainer>,
+    repr: Box<Repr>,
 }
 
 impl Error {
     /// Returns the error kind.
     pub fn kind(&self) -> ErrorKind {
-        self.inner.kind()
+        self.repr.error.kind()
     }
 
     /// Returns a reference to the inner error container.
     pub fn get_ref(&self) -> &ErrorContainer {
-        &self.inner
+        &self.repr.error
     }
 
     /// Returns a reference to the inner error if the type matches.
     pub fn downcast_ref<T: 'static + error::Error>(&self) -> Option<&T> {
-        self.inner.as_error().downcast_ref::<T>()
+        self.repr.error.as_error().downcast_ref::<T>()
+    }
+
+    /// Returns the syntactic position if available.
+    pub fn position(&self) -> Option<&SyntacticPosition> {
+        self.repr.position.as_ref()
+    }
+
+    /// Creates a new `error` with the given syntactic position info.
+    pub(crate) fn with_position(error: ErrorContainer, position: SyntacticPosition) -> Self {
+        Self {
+            repr: Box::new(Repr::with_position(error, position)),
+        }
+    }
+
+    /// Sets the syntactic position and returns the new error.
+    pub(crate) fn and_position(mut self, position: SyntacticPosition) -> Self {
+        self.repr.position = Some(position);
+        self
     }
 }
 
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        self.inner.fmt(f)
+        self.repr.error.fmt(f)
     }
 }
 
 impl error::Error for Error {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        self.inner.source()
+        self.repr.error.source()
     }
 }
 
@@ -59,7 +79,34 @@ where
 {
     fn from(e: T) -> Self {
         Error {
-            inner: Box::new(e.into()),
+            repr: Box::new(Repr::new(e.into())),
+        }
+    }
+}
+
+/// Internal representation of parsing error.
+#[derive(Debug)]
+struct Repr {
+    /// Error.
+    error: ErrorContainer,
+    /// Syntactic position.
+    position: Option<SyntacticPosition>,
+}
+
+impl Repr {
+    /// Creates a new `Repr`.
+    pub(crate) fn new(error: ErrorContainer) -> Self {
+        Self {
+            error,
+            position: None,
+        }
+    }
+
+    /// Creates a new `Repr` with the given syntactic position info.
+    pub(crate) fn with_position(error: ErrorContainer, position: SyntacticPosition) -> Self {
+        Self {
+            error,
+            position: Some(position),
         }
     }
 }