+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 # Change Log
 
 ## [Unreleased]
+* Syntactic position information is supported.
+  Syntactic position contains node path, node index, attribute index, etc.
+  This will make errors and warnings more detailed and useful.
+
+### Breaking changes
+* `pull_parser::v7400::Parser::set_warning_handler()` now requires
+  `'static + FnMut(Warning, &SyntacticPosition) -> Result<()>` as warning
+  hander (note that `&SyntacticPosition` argument is added).
+    + By this change, warning handler can use position information where the
+      warning happened.
+
 ### Added
+* `pull_parser::SyntacticPosition` is added.
+* `pull_parser::error::Error::position()` is added.
 * `pull_parser::v7400::Parser::skip_current_node()` is added.
 
 ## [0.1.0]